internal/agent: rename agentdiscovery to agent and clean up names
