+++ internal/agent/discovery_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2022 Datadog, Inc.
 
-package agentdiscovery_test
+package agent_test
 
 import (
 	"fmt"
@@ -16,14 +16,14 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"gopkg.in/DataDog/dd-trace-go.v1/internal/agentdiscovery"
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/agent"
 )
 
 func TestHTTPClient(t *testing.T) {
 	t.Run("default", func(t *testing.T) {
 		server := httptest.NewServer(http.DefaultServeMux)
 		defer server.Close()
-		client := agentdiscovery.HTTPClient()
+		client := agent.HTTPClient()
 		r, err := client.Head(server.URL)
 		if err != nil {
 			t.Error(err)
@@ -43,7 +43,7 @@ func TestHTTPClient(t *testing.T) {
 		defer server.Close()
 
 		t.Setenv("DD_TRACE_AGENT_URL", fmt.Sprintf("unix://%s", udspath))
-		client := agentdiscovery.HTTPClient()
+		client := agent.HTTPClient()
 		// TODO: can we be sure the server is actually *serving* here?
 		// We started it in a separate goroutine so there's no guarantee
 		// that between then and now that the goroutine was ever
@@ -60,20 +60,20 @@ func TestResolveAgentAddr(t *testing.T) {
 	t.Run("default", func(t *testing.T) {
 		unsetenv(t, "DD_AGENT_HOST")
 		unsetenv(t, "DD_TRACE_AGENT_PORT")
-		addr := agentdiscovery.ResolveAgentAddr()
+		addr := agent.ResolveAgentAddr()
 		assert.Equal(t, "localhost:8126", addr)
 	})
 
 	t.Run("host-port-env", func(t *testing.T) {
 		t.Setenv("DD_AGENT_HOST", "foobar")
 		t.Setenv("DD_TRACE_AGENT_PORT", "9999")
-		addr := agentdiscovery.ResolveAgentAddr()
+		addr := agent.ResolveAgentAddr()
 		assert.Equal(t, "foobar:9999", addr)
 	})
 
 	t.Run("url-env", func(t *testing.T) {
 		t.Setenv("DD_TRACE_AGENT_URL", "http://datadogagent.example.com")
-		addr := agentdiscovery.ResolveAgentAddr()
+		addr := agent.ResolveAgentAddr()
 		assert.Equal(t, "datadogagent.example.com", addr)
 	})
 
@@ -81,7 +81,7 @@ func TestResolveAgentAddr(t *testing.T) {
 		unsetenv(t, "DD_AGENT_HOST")
 		unsetenv(t, "DD_TRACE_AGENT_PORT")
 		t.Setenv("DD_TRACE_AGENT_URL", "/var/file/socket.sock")
-		addr := agentdiscovery.ResolveAgentAddr()
+		addr := agent.ResolveAgentAddr()
 		// When the connecting via a Unix domain socket, the address
 		// doesn't actually matter
 		assert.Equal(t, "localhost:8126", addr)
@@ -97,11 +97,11 @@ func TestFeatures(t *testing.T) {
 	}))
 	defer server.Close()
 
-	features, err := agentdiscovery.AgentFeatures(server.Listener.Addr().String(), new(http.Client))
+	features, err := agent.LoadFeatures(server.Listener.Addr().String(), new(http.Client))
 	if err != nil {
 		t.Fatal(err)
 	}
-	expected := &agentdiscovery.Features{
+	expected := &agent.Features{
 		Endpoints:     []string{"/v0.6/stats"},
 		ClientDropP0s: true,
 		StatsdPort:    8999,
